use long

+++ src/main/java/com/example/budge/homework/Period.java
@@ -12,7 +12,7 @@ public class Period {
     private final LocalDate start;
     private final LocalDate end;
 
-    public int getOverlappingDays(Period another) {
+    public long getOverlappingDays(Period another) {
 
         if (this.start.isAfter(another.getEnd()) || this.end.isBefore(another.getStart())) {
             return 0;
@@ -21,6 +21,6 @@ public class Period {
         LocalDate periodStart = this.start.isAfter(another.getStart()) ? this.start : another.getStart();
         LocalDate periodEnd = this.end.isBefore(another.getEnd()) ? this.end : another.getEnd();
 
-        return (int) ChronoUnit.DAYS.between(periodStart, periodEnd) + 1;
+        return ChronoUnit.DAYS.between(periodStart, periodEnd) + 1L;
     }
 }
